Output data printing improved
