+++ src/main/java/com/github/bartoszreszka/lighting_chart/Main.java
@@ -1,6 +1,6 @@
 package com.github.bartoszreszka.lighting_chart;
 
-import com.github.bartoszreszka.lighting_chart.view.start.GUI;
+import com.github.bartoszreszka.lighting_chart.controller.GUI;
 
 /**
  * Program calculates and generates "lighting chart" for following phenomenons in a given month:
@@ -13,7 +13,7 @@ import com.github.bartoszreszka.lighting_chart.view.start.GUI;
  * </ul>
  * <br>
  * @see     <a href="https://shredzone.org/">commons-suncalc</a>;
- * @version 1.0.0
+ * @version 1.0.1
  * @author  Bartosz Reszka
  * */
 public class Main {
