+++ src/main/java/com/github/bartoszreszka/lighting_chart/controller/GUI.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.view.start.GUI">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.controller.GUI">
   <grid id="cbd77" binding="contentPane" layout-manager="GridLayoutManager" row-count="4" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="10" left="10" bottom="10" right="10"/>
     <constraints>
@@ -64,13 +64,16 @@
             </constraints>
             <properties>
               <enabled value="true"/>
+              <toolTipText value="Wybierz miesiąc."/>
             </properties>
           </component>
           <component id="9a131" class="javax.swing.JSpinner" binding="spinner2" custom-create="true" default-binding="true">
             <constraints>
               <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
             </constraints>
-            <properties/>
+            <properties>
+              <toolTipText value="Wybierz rok."/>
+            </properties>
           </component>
         </children>
       </grid>
@@ -90,7 +93,7 @@
             </constraints>
             <properties>
               <model/>
-              <toolTipText value="Wybierz port lub uzupełnij współrzędne"/>
+              <toolTipText value="Wybierz port lub wpisz własne współrzędne poniżej."/>
             </properties>
           </component>
         </children>
@@ -127,7 +130,7 @@
             </constraints>
             <properties>
               <text value=""/>
-              <toolTipText value="Wprowadź wartość kątową (od -90 do 90) z wartością dziesiętną po kropce lub stopnie i minuty oddzielnie"/>
+              <toolTipText value="Wprowadź wartość kątową (od -90 do 90) z wartością dziesiętną po kropce."/>
             </properties>
           </component>
           <component id="1de21" class="javax.swing.JTextField" binding="longDegField">
@@ -137,7 +140,7 @@
               </grid>
             </constraints>
             <properties>
-              <toolTipText value="Wprowadź wartość kątową (od -180 do 180) z wartością dziesiętną po kropce lub stopnie i minuty oddzielnie"/>
+              <toolTipText value="Wprowadź wartość kątową (od -180 do 180) z wartością dziesiętną po kropce."/>
             </properties>
           </component>
           <component id="4699" class="javax.swing.JTextField" binding="latMinutesField">
@@ -147,8 +150,10 @@
               </grid>
             </constraints>
             <properties>
+              <enabled value="false"/>
               <text value=""/>
               <toolTipText value="Minuty kątowe"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="d2e1d" class="javax.swing.JTextField" binding="longMinutesField">
@@ -158,7 +163,9 @@
               </grid>
             </constraints>
             <properties>
+              <enabled value="false"/>
               <toolTipText value="Minuty kątowe"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="c4c6c" class="javax.swing.JLabel">
@@ -167,6 +174,7 @@
             </constraints>
             <properties>
               <text value="deg"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="8d966" class="javax.swing.JLabel">
@@ -175,6 +183,7 @@
             </constraints>
             <properties>
               <text value="deg"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="21cd8" class="javax.swing.JLabel">
@@ -183,6 +192,7 @@
             </constraints>
             <properties>
               <text value="'"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="c078e" class="javax.swing.JLabel">
@@ -191,6 +201,7 @@
             </constraints>
             <properties>
               <text value="'"/>
+              <visible value="false"/>
             </properties>
           </component>
           <component id="8e5fd" class="javax.swing.JLabel">
@@ -210,6 +221,7 @@
             <properties>
               <focusable value="true"/>
               <text value=""/>
+              <toolTipText value="Wprowadź nazwę punktu."/>
             </properties>
           </component>
         </children>
