+++ src/main/java/com/github/bartoszreszka/lighting_chart/view/Chart.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.view.chart.Chart">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.view.Chart">
   <grid id="cbd77" binding="contentPane" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="48" y="54" width="1756" height="702"/>
