+++ src/main/java/com/github/bartoszreszka/lighting_chart/view/HoursPane.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.view.chart.SunrisePane">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.github.bartoszreszka.lighting_chart.view.HoursPane">
   <grid id="27dc6" row-count="1" column-count="1" layout-manager="CardLayout">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
