+++ src/main/java/com/github/bartoszreszka/lighting_chart/view/MoonphasePane.java
@@ -1,7 +1,7 @@
-package com.github.bartoszreszka.lighting_chart.view.chart;
+package com.github.bartoszreszka.lighting_chart.view;
 
 import com.github.bartoszreszka.lighting_chart.model.APane;
-import com.github.bartoszreszka.lighting_chart.controller.Computations;
+import com.github.bartoszreszka.lighting_chart.model.Computations;
 import com.github.bartoszreszka.lighting_chart.model.Month;
 import org.shredzone.commons.suncalc.MoonPhase;
