+++ src/main/java/com/github/bartoszreszka/lighting_chart/view/MoonrisePane.java
@@ -1,4 +1,4 @@
-package com.github.bartoszreszka.lighting_chart.view.chart;
+package com.github.bartoszreszka.lighting_chart.view;
 
 import com.github.bartoszreszka.lighting_chart.model.APane;
 import com.github.bartoszreszka.lighting_chart.model.Day;
@@ -7,9 +7,9 @@ import com.github.bartoszreszka.lighting_chart.model.Phenomena;
 import java.awt.*;
 import java.util.Formatter;
 
-import static com.github.bartoszreszka.lighting_chart.controller.Computations.*;
+import static com.github.bartoszreszka.lighting_chart.model.Computations.*;
 
-public class MoonsetPane extends APane {
+public class MoonrisePane extends APane {
     @Override
     protected void drawPane(Graphics g) {
         int i = 0;
@@ -17,8 +17,8 @@ public class MoonsetPane extends APane {
             StringBuilder sb = new StringBuilder();
             Formatter f = new Formatter(sb);
             f.format("%02d:%02d",
-                    getHourOf(Phenomena.MOONSET, day),
-                    getMinutesOf(Phenomena.MOONSET, day));
+                    getHourOf(Phenomena.MOONRISE, day),
+                    getMinutesOf(Phenomena.MOONRISE, day));
             g.drawString(String.valueOf(f),
                     0,
                     i * Chart.dayHeightInPixels + (Chart.fontHeight / 2));
